add a basic flask-restless API (#29)
This is straight out of the restless quickstart. Up-to-date
restless packages are in updates-testing at present, you should
probably test with those, not the ancient ones in stable that
probably don't work. This is untested, as I wasn't able to get
a dev instance of autocloud running.

If it's working, something like:
http://foo/autocloud/api/job_details
should work (I think it should return a JSON-formatted dict of
all jobs). /job_details/jobid should return a single job.